Split medication tests by properties and methods
